refactor: use new spending condition interface for challenging output exits from in-flight transactions

+++ plasma_framework/contracts/src/exits/payment/spendingConditions/IPaymentSpendingCondition.sol
@@ -15,7 +15,7 @@ interface IPaymentSpendingCondition {
         uint256 _utxoPos,
         bytes32 _outputId,
         bytes calldata _spendingTx,
-        uint8 _inputIndex,
+        uint16 _inputIndex,
         bytes calldata _witness
     ) external view returns (bool);
 }
